// assets/react/types/dossier.types.ts
export interface DossiersInterface {
  readonly aanmaakDatumTijd: string
  readonly allegroNummer: string
  readonly clientNaam: string
  readonly clientVoorletters: string
  readonly id: string
  readonly isInPrullenbak: boolean
  readonly organisatie: string
  readonly regasNummer: string
  readonly status: string
}